Finished Filtering Scripts (among other things)
Finished scripts that help
(1) calculating the age of each user,
(2) removing those less than 18,
(3) removing those of non-Filipino nationality,
(4) merge personality data and prescense of social media account into one dataframe, and
(5) saves a final list of participants

Outside of what files are commited, file/folder structure of the local was organized. Data and Constant folders are populated, but will not be synced.