Created scripts to extract demographics of the users
Includes:
(1) total participant counts
(2) counts for sex categories
(3) stats for age + age groups
(4) counts for nationality

This is done across 4 groups defined by ownership of a social media account.

Demographics are saved to a CSV file.

Not sure if thre is a way to optimize this... But tbh, I don't think I'll be running this too much nor do i think this is that memory/computation expensive.